requirements.txt
PyQt5
mss
pillow
pynput
pytesseract
